requirements.txt
torch~=2.7.0
numpy~=2.2.5
pandas~=2.2.3
matplotlib~=3.10.1
transformers~=4.51.3
pillow~=11.2.1
torchvision~=0.22.0
seaborn~=0.13.2